Organized codemods folder. Fixed some warnings
